#107: add update fonts and theme pages
