+++ lib/catalog/catalog_settings.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:noted_app/ui/common/icon/noted_icons.dart';
 import 'package:noted_app/ui/common/layout/noted_header_page.dart';
 import 'package:noted_app/ui/settings/settings_row.dart';
-import 'package:noted_app/ui/settings/style/color_switcher.dart';
-import 'package:noted_app/ui/settings/style/font_switcher.dart';
+import 'package:noted_app/ui/settings/style/update_fonts_page.dart';
+import 'package:noted_app/ui/settings/style/update_theme_page.dart';
 
 class CatalogSettings extends StatelessWidget {
   const CatalogSettings({super.key});
@@ -37,10 +37,10 @@ class CatalogSettings extends StatelessWidget {
   }
 
   void _navigateToColors(BuildContext context) {
-    Navigator.of(context).push(MaterialPageRoute(builder: (context) => const ColorSwitcher()));
+    Navigator.of(context).push(MaterialPageRoute(builder: (context) => const UpdateThemePage()));
   }
 
   void _navigateToFonts(BuildContext context) {
-    Navigator.of(context).push(MaterialPageRoute(builder: (context) => const FontSwitcher()));
+    Navigator.of(context).push(MaterialPageRoute(builder: (context) => const UpdateFontsPage()));
   }
 }
