+++ lib/repository/auth/local_auth_repository.dart
@@ -41,6 +41,12 @@ class LocalAuthRepository extends AuthRepository implements Disposable {
   @override
   Stream<NotedUser> get userStream => _userStreamController.stream;
 
+  LocalAuthRepository({NotedUser user = const NotedUser.empty()}) {
+    if (user.isNotEmpty) {
+      _currentUser = user;
+    }
+  }
+
   @override
   Future<void> createUserWithEmailAndPassword({String email = '', String password = ''}) async {
     await Future.delayed(Duration(milliseconds: _msDelay));
